pyproject.toml
[tool.poetry]
name = "line-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
line-bot-sdk = "^3.13.0"
cohere = "^5.11.0"
setuptools = "^75.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
